+++ egov/egov-egf/src/main/java/org/egov/services/report/FunctionwiseIEService.java
@@ -1051,7 +1051,7 @@ public class FunctionwiseIEService
                         .addScalar("deptName", StringType.INSTANCE)
                         .setString("isBeRe", isBeRe)
                         .setLong("finYearId", reportSearch.getFinYearId())
-                        .setInteger("fundId", reportSearch.getFund().getId())
+                        .setInteger("fundId", reportSearch.getFund().getId().intValue())
                         .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
 
                 if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
@@ -1071,7 +1071,7 @@ public class FunctionwiseIEService
                         .addScalar("isMajor", BooleanType.INSTANCE)
                         .setString("isBeRe", isBeRe)
                         .setLong("finYearId", reportSearch.getFinYearId())
-                        .setInteger("fundId", reportSearch.getFund().getId())
+                        .setInteger("fundId", reportSearch.getFund().getId().intValue())
                         .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
                 if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
                         && reportSearch.getFunction().getId() != -1)
@@ -1088,7 +1088,7 @@ public class FunctionwiseIEService
                     .setString("isBeRe", isBeRe)
                     .setDate("asOnDate", reportSearch.getAsOnDate())
                     .setLong("finYearId", reportSearch.getFinYearId())
-                    .setInteger("fundId", reportSearch.getFund().getId())
+                    .setInteger("fundId", reportSearch.getFund().getId().intValue())
                     .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
 
             if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
@@ -1109,7 +1109,7 @@ public class FunctionwiseIEService
                     .setString("isBeRe", isBeRe)
                     .setDate("asOnDate", reportSearch.getAsOnDate())
                     .setLong("finYearId", reportSearch.getFinYearId())
-                    .setInteger("fundId", reportSearch.getFund().getId())
+                    .setInteger("fundId", reportSearch.getFund().getId().intValue())
                     .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
             if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
                     && reportSearch.getFunction().getId() != -1)
@@ -1136,7 +1136,7 @@ public class FunctionwiseIEService
                     .setString("isBeRe", isBeRe)
                     .setDate("asOnDate", reportSearch.getAsOnDate())
                     .setLong("finYearId", reportSearch.getFinYearId())
-                    .setInteger("fundId", reportSearch.getFund().getId())
+                    .setInteger("fundId", reportSearch.getFund().getId().intValue())
 
                     .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
             if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
@@ -1157,7 +1157,7 @@ public class FunctionwiseIEService
                     .setString("isBeRe", isBeRe)
                     .setDate("asOnDate", reportSearch.getAsOnDate())
                     .setLong("finYearId", reportSearch.getFinYearId())
-                    .setInteger("fundId", reportSearch.getFund().getId())
+                    .setInteger("fundId", reportSearch.getFund().getId().intValue())
                     .setResultTransformer(Transformers.aliasToBean(CommonReportBean.class));
             if (reportSearch.getFunction() != null && reportSearch.getFunction().getId() != null
                     && reportSearch.getFunction().getId() != -1)
